--- packages/skia/src/renderer/Canvas2.web.tsx
@@ -1,6 +1,0 @@
-import type { CanvasProps } from "./Canvas";
-import { Canvas } from "./Canvas";
-
-export type Canvas2Props = CanvasProps;
-
-export const Canvas2 = Canvas;
